ref(nav): Move issue navigation into secondary/sections/issues
