+++ static/app/views/dashboards/navigation.tsx
@@ -6,7 +6,7 @@ import useOrganization from 'sentry/utils/useOrganization';
 import type {DashboardListItem} from 'sentry/views/dashboards/types';
 import {NAV_GROUP_LABELS} from 'sentry/views/nav/constants';
 import {usePrefersStackedNav} from 'sentry/views/nav/prefersStackedNav';
-import {SecondaryNav} from 'sentry/views/nav/secondary';
+import {SecondaryNav} from 'sentry/views/nav/secondary/secondary';
 import {PrimaryNavGroup} from 'sentry/views/nav/types';
 
 type DashboardsNavigationProps = {
