+++ static/app/views/insights/navigation.tsx
@@ -29,7 +29,7 @@ import {DOMAIN_VIEW_BASE_URL} from 'sentry/views/insights/pages/settings';
 import {NAV_GROUP_LABELS} from 'sentry/views/nav/constants';
 import {usePrefersStackedNav} from 'sentry/views/nav/prefersStackedNav';
 import ProjectIcon from 'sentry/views/nav/projectIcon';
-import {SecondaryNav} from 'sentry/views/nav/secondary';
+import {SecondaryNav} from 'sentry/views/nav/secondary/secondary';
 import {PrimaryNavGroup} from 'sentry/views/nav/types';
 import {isLinkActive} from 'sentry/views/nav/utils';
