+++ static/app/views/nav/secondary/sections/issues/issueViews/issueViewNavQueryCount.tsx
