+++ static/app/views/nav/secondary/sections/issues/issueViews/useUpdateGroupSearchViewLastVisited.tsx
