+++ static/app/views/nav/secondary/sections/issues/issuesSecondaryNav.tsx
@@ -8,10 +8,10 @@ import useOrganization from 'sentry/utils/useOrganization';
 import {useParams} from 'sentry/utils/useParams';
 import type {IssueView} from 'sentry/views/issueList/issueViews/issueViews';
 import {useFetchGroupSearchViews} from 'sentry/views/issueList/queries/useFetchGroupSearchViews';
-import {IssueViewNavItems} from 'sentry/views/nav/issueViews/issueViewNavItems';
-import {useUpdateGroupSearchViewLastVisited} from 'sentry/views/nav/issueViews/useUpdateGroupSearchViewLastVisited';
 import {usePrefersStackedNav} from 'sentry/views/nav/prefersStackedNav';
-import {SecondaryNav} from 'sentry/views/nav/secondary';
+import {SecondaryNav} from 'sentry/views/nav/secondary/secondary';
+import {IssueViewNavItems} from 'sentry/views/nav/secondary/sections/issues/issueViews/issueViewNavItems';
+import {useUpdateGroupSearchViewLastVisited} from 'sentry/views/nav/secondary/sections/issues/issueViews/useUpdateGroupSearchViewLastVisited';
 import {PrimaryNavGroup} from 'sentry/views/nav/types';
 
 interface IssuesWrapperProps extends RouteComponentProps {
