+++ static/app/views/pipeline/navigation.tsx
@@ -3,7 +3,7 @@ import {Fragment} from 'react';
 import {t} from 'sentry/locale';
 import useOrganization from 'sentry/utils/useOrganization';
 import {NAV_GROUP_LABELS} from 'sentry/views/nav/constants';
-import {SecondaryNav} from 'sentry/views/nav/secondary';
+import {SecondaryNav} from 'sentry/views/nav/secondary/secondary';
 import {PrimaryNavGroup} from 'sentry/views/nav/types';
 import {makePipelinePathname} from 'sentry/views/pipeline/pathnames';
 import {COVERAGE_BASE_URL, TESTS_BASE_URL} from 'sentry/views/pipeline/settings';
