+++ static/app/views/settings/components/settingsNavItem.tsx
@@ -8,7 +8,7 @@ import HookOrDefault from 'sentry/components/hookOrDefault';
 import {Tooltip} from 'sentry/components/tooltip';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {SecondaryNav} from 'sentry/views/nav/secondary';
+import {SecondaryNav} from 'sentry/views/nav/secondary/secondary';
 
 type Props = {
   label: React.ReactNode;
